requirements.txt
obs-websocket-py
python-dotenv
flask